This mod adds two interiors to GTA V that were left in the game, with modifications to make them load correctly.
The mod adds two interiors that Rockstar left over in the game files from the beta version of the game/GTA Online updates. Additional code and modifications were made to get the interiors to load in the game.
-The hidden High Life Update Garage: Located in Vinewood Hills
-The beta version of the FIB interior: Located on the 7th floor of the FIB Building in Downtown Los Santos

The mod does not replace any interiors. They are properly spawned via a DLC ymap/ytyp map and have MLO characteristics.